feat: login and register user form completed and working

+++ frontend/src/pages/RegisterForm.jsx
@@ -1,6 +1,7 @@
 import { useState } from "react";
 import { Link } from 'react-router-dom';
-import { loginUser } from "../services/userService";
+import { loginUser, registerUser } from "../services/userService";
+import { useNavigate } from "react-router-dom";
 
 function RegisterForm({ setUser }) {
 
@@ -8,14 +9,19 @@ function RegisterForm({ setUser }) {
 
   const [email, setEmail] = useState("")
   const [password, setPassword] = useState("")
+  const [confirmPassword, setConfirmPassword] = useState("")
 
   const [firstName, setFirstName] = useState("")
   const [lastName, setLastName] = useState("")
-  const [role, setRole] = useState("GUARDIAN");
+  const [role, setRole] = useState("GUARDIAN")
 
-  const [error, setError] = useState(null);
+  const [error, setError] = useState(null)
+
+  const navigate = useNavigate()
 
   function handleChange(input) {
+    setError(null);
+
     const { name, value } = input.target
     switch (name) {
       case "email":
@@ -23,6 +29,9 @@ function RegisterForm({ setUser }) {
         break;
       case "password":
         setPassword(value);
+        break;
+      case "confirmPassword":
+        setConfirmPassword(value);
         break;
       case "firstName":
         setFirstName(value);
@@ -38,21 +47,35 @@ function RegisterForm({ setUser }) {
     }
   }
 
-  function handleSubmit(form) {
-    
+  async function handleSubmit(form) {
     form.preventDefault()
+
     if (isRegister) {
-      console.log("Registrando:", { email, password, firstName, lastName });
-      setUser({ username: firstName, role, loggedIn: true });
-      navigate("/map");
+      if (confirmPassword !== password) {
+        setError("Las contrasenas no coinciden.")
+
+      } else {
+        try {
+          const data = await registerUser( firstName, lastName, email, password, role );
+          localStorage.setItem("token", data.token);
+          setUser({ username: firstName, role: role, loggedIn: true });
+          navigate("/map");
+
+        } catch (err) {
+          setError(err.message)
+        }
+      }
 
     } else {
 
       try {
-        const data = loginUser(email, password);
-        localStorage.setItem("token", data.token);      // guardar JWT
-        setUser({ username: data.username, role: data.role, loggedIn: true });
+        const data = await loginUser(email, password);
+        localStorage.setItem("token", data.token);
+        console.log(data);
+
+        setUser({ username: data.firstName, role: data.role, loggedIn: true });
         navigate("/map");
+
       } catch (err) {
         setError(err.message);
       }
@@ -63,7 +86,7 @@ function RegisterForm({ setUser }) {
 
     <div>
 
-      <h1>{isRegister ? "REGISTRO" : "LOGIN"}</h1>
+      <h1>{isRegister ? "REGISTRO" : "LOG IN"}</h1>
 
       <form onSubmit={handleSubmit}>
         {isRegister && (
@@ -123,10 +146,10 @@ function RegisterForm({ setUser }) {
 
         {isRegister && (
           <input
-            type="confirm-password"
+            type="password"
             placeholder="Confirmar contrasena"
-            name="password"
-            value={password}
+            name="confirmPassword"
+            value={confirmPassword}
             onChange={handleChange}
             required
           />)}
@@ -136,10 +159,16 @@ function RegisterForm({ setUser }) {
         <button type="submit">{isRegister ? "REGÍSTRATE" : "LOGIN"}</button>
       </form>
 
+      {error && (
+        <div className="error-message">
+          {error}
+        </div>
+      )}
+
       <br />
 
       <p>
-        {isRegister ? "¿Ya tienes cuenta?" : "¿No tienes cuenta?"}{" "}
+        {isRegister ? "¿Ya tienes cuenta?" : "¿No tienes cuenta?"}
         <button onClick={() => setIsRegister(!isRegister)}>
           {isRegister ? "INICIA SESIÓN" : "REGÍSTRATE"}
         </button>
